+++ internal/provider/resource_iosxr_router_static_ipv4_unicast.go
@@ -114,6 +114,25 @@ func (r *RouterStaticIPv4UnicastResource) Schema(ctx context.Context, req resour
 								int64validator.Between(1, 254),
 							},
 						},
+						"permanent": schema.BoolAttribute{
+							MarkdownDescription: helpers.NewAttributeDescription("Permanent route").String,
+							Optional:            true,
+						},
+						"track": schema.StringAttribute{
+							MarkdownDescription: helpers.NewAttributeDescription("Enable object tracking for static route").String,
+							Optional:            true,
+							Validators: []validator.String{
+								stringvalidator.LengthBetween(1, 32),
+								stringvalidator.RegexMatches(regexp.MustCompile(`[\w\-\.:,_@#%$\+=\|;]+`), ""),
+							},
+						},
+						"metric": schema.Int64Attribute{
+							MarkdownDescription: helpers.NewAttributeDescription("Set metric for this route").AddIntegerRangeDescription(1, 16777214).String,
+							Optional:            true,
+							Validators: []validator.Int64{
+								int64validator.Between(1, 16777214),
+							},
+						},
 					},
 				},
 			},
@@ -158,6 +177,25 @@ func (r *RouterStaticIPv4UnicastResource) Schema(ctx context.Context, req resour
 								int64validator.Between(1, 254),
 							},
 						},
+						"permanent": schema.BoolAttribute{
+							MarkdownDescription: helpers.NewAttributeDescription("Permanent route").String,
+							Optional:            true,
+						},
+						"track": schema.StringAttribute{
+							MarkdownDescription: helpers.NewAttributeDescription("Enable object tracking for static route").String,
+							Optional:            true,
+							Validators: []validator.String{
+								stringvalidator.LengthBetween(1, 32),
+								stringvalidator.RegexMatches(regexp.MustCompile(`[\w\-\.:,_@#%$\+=\|;]+`), ""),
+							},
+						},
+						"metric": schema.Int64Attribute{
+							MarkdownDescription: helpers.NewAttributeDescription("Set metric for this route").AddIntegerRangeDescription(1, 16777214).String,
+							Optional:            true,
+							Validators: []validator.Int64{
+								int64validator.Between(1, 16777214),
+							},
+						},
 					},
 				},
 			},
@@ -193,6 +231,215 @@ func (r *RouterStaticIPv4UnicastResource) Schema(ctx context.Context, req resour
 							Optional:            true,
 							Validators: []validator.Int64{
 								int64validator.Between(1, 254),
+							},
+						},
+						"permanent": schema.BoolAttribute{
+							MarkdownDescription: helpers.NewAttributeDescription("Permanent route").String,
+							Optional:            true,
+						},
+						"track": schema.StringAttribute{
+							MarkdownDescription: helpers.NewAttributeDescription("Enable object tracking for static route").String,
+							Optional:            true,
+							Validators: []validator.String{
+								stringvalidator.LengthBetween(1, 32),
+								stringvalidator.RegexMatches(regexp.MustCompile(`[\w\-\.:,_@#%$\+=\|;]+`), ""),
+							},
+						},
+						"metric": schema.Int64Attribute{
+							MarkdownDescription: helpers.NewAttributeDescription("Set metric for this route").AddIntegerRangeDescription(1, 16777214).String,
+							Optional:            true,
+							Validators: []validator.Int64{
+								int64validator.Between(1, 16777214),
+							},
+						},
+					},
+				},
+			},
+			"vrfs": schema.ListNestedAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Destination VRF").String,
+				Optional:            true,
+				NestedObject: schema.NestedAttributeObject{
+					Attributes: map[string]schema.Attribute{
+						"vrf_name": schema.StringAttribute{
+							MarkdownDescription: helpers.NewAttributeDescription("Destination VRF").String,
+							Required:            true,
+							Validators: []validator.String{
+								stringvalidator.LengthBetween(1, 32),
+								stringvalidator.RegexMatches(regexp.MustCompile(`[\w\-\.:,_@#%$\+=\|;]+`), ""),
+							},
+						},
+						"nexthop_interfaces": schema.ListNestedAttribute{
+							MarkdownDescription: helpers.NewAttributeDescription("Forwarding interface").String,
+							Optional:            true,
+							NestedObject: schema.NestedAttributeObject{
+								Attributes: map[string]schema.Attribute{
+									"interface_name": schema.StringAttribute{
+										MarkdownDescription: helpers.NewAttributeDescription("Forwarding interface").String,
+										Required:            true,
+										Validators: []validator.String{
+											stringvalidator.RegexMatches(regexp.MustCompile(`[a-zA-Z0-9.:_/-]+`), ""),
+										},
+									},
+									"description": schema.StringAttribute{
+										MarkdownDescription: helpers.NewAttributeDescription("description of the static route").String,
+										Optional:            true,
+										Validators: []validator.String{
+											stringvalidator.LengthBetween(1, 30),
+										},
+									},
+									"tag": schema.Int64Attribute{
+										MarkdownDescription: helpers.NewAttributeDescription("Set tag for this route").AddIntegerRangeDescription(1, 4294967295).String,
+										Optional:            true,
+										Validators: []validator.Int64{
+											int64validator.Between(1, 4294967295),
+										},
+									},
+									"distance_metric": schema.Int64Attribute{
+										MarkdownDescription: helpers.NewAttributeDescription("Distance metric for this route").AddIntegerRangeDescription(1, 254).String,
+										Optional:            true,
+										Validators: []validator.Int64{
+											int64validator.Between(1, 254),
+										},
+									},
+									"permanent": schema.BoolAttribute{
+										MarkdownDescription: helpers.NewAttributeDescription("Permanent route").String,
+										Optional:            true,
+									},
+									"track": schema.StringAttribute{
+										MarkdownDescription: helpers.NewAttributeDescription("Enable object tracking for static route").String,
+										Optional:            true,
+										Validators: []validator.String{
+											stringvalidator.LengthBetween(1, 32),
+											stringvalidator.RegexMatches(regexp.MustCompile(`[\w\-\.:,_@#%$\+=\|;]+`), ""),
+										},
+									},
+									"metric": schema.Int64Attribute{
+										MarkdownDescription: helpers.NewAttributeDescription("Set metric for this route").AddIntegerRangeDescription(1, 16777214).String,
+										Optional:            true,
+										Validators: []validator.Int64{
+											int64validator.Between(1, 16777214),
+										},
+									},
+								},
+							},
+						},
+						"nexthop_interface_addresses": schema.ListNestedAttribute{
+							MarkdownDescription: helpers.NewAttributeDescription("Forwarding interface").String,
+							Optional:            true,
+							NestedObject: schema.NestedAttributeObject{
+								Attributes: map[string]schema.Attribute{
+									"interface_name": schema.StringAttribute{
+										MarkdownDescription: helpers.NewAttributeDescription("Forwarding interface").String,
+										Required:            true,
+										Validators: []validator.String{
+											stringvalidator.RegexMatches(regexp.MustCompile(`[a-zA-Z0-9.:_/-]+`), ""),
+										},
+									},
+									"address": schema.StringAttribute{
+										MarkdownDescription: helpers.NewAttributeDescription("Forwarding router's address").String,
+										Required:            true,
+										Validators: []validator.String{
+											stringvalidator.RegexMatches(regexp.MustCompile(`(([0-9]|[1-9][0-9]|1[0-9][0-9]|2[0-4][0-9]|25[0-5])\.){3}([0-9]|[1-9][0-9]|1[0-9][0-9]|2[0-4][0-9]|25[0-5])(%[\p{N}\p{L}]+)?`), ""),
+											stringvalidator.RegexMatches(regexp.MustCompile(`[0-9\.]*`), ""),
+										},
+									},
+									"description": schema.StringAttribute{
+										MarkdownDescription: helpers.NewAttributeDescription("description of the static route").String,
+										Optional:            true,
+										Validators: []validator.String{
+											stringvalidator.LengthBetween(1, 30),
+										},
+									},
+									"tag": schema.Int64Attribute{
+										MarkdownDescription: helpers.NewAttributeDescription("Set tag for this route").AddIntegerRangeDescription(1, 4294967295).String,
+										Optional:            true,
+										Validators: []validator.Int64{
+											int64validator.Between(1, 4294967295),
+										},
+									},
+									"distance_metric": schema.Int64Attribute{
+										MarkdownDescription: helpers.NewAttributeDescription("Distance metric for this route").AddIntegerRangeDescription(1, 254).String,
+										Optional:            true,
+										Validators: []validator.Int64{
+											int64validator.Between(1, 254),
+										},
+									},
+									"permanent": schema.BoolAttribute{
+										MarkdownDescription: helpers.NewAttributeDescription("Permanent route").String,
+										Optional:            true,
+									},
+									"track": schema.StringAttribute{
+										MarkdownDescription: helpers.NewAttributeDescription("Enable object tracking for static route").String,
+										Optional:            true,
+										Validators: []validator.String{
+											stringvalidator.LengthBetween(1, 32),
+											stringvalidator.RegexMatches(regexp.MustCompile(`[\w\-\.:,_@#%$\+=\|;]+`), ""),
+										},
+									},
+									"metric": schema.Int64Attribute{
+										MarkdownDescription: helpers.NewAttributeDescription("Set metric for this route").AddIntegerRangeDescription(1, 16777214).String,
+										Optional:            true,
+										Validators: []validator.Int64{
+											int64validator.Between(1, 16777214),
+										},
+									},
+								},
+							},
+						},
+						"nexthop_addresses": schema.ListNestedAttribute{
+							MarkdownDescription: helpers.NewAttributeDescription("Forwarding router's address").String,
+							Optional:            true,
+							NestedObject: schema.NestedAttributeObject{
+								Attributes: map[string]schema.Attribute{
+									"address": schema.StringAttribute{
+										MarkdownDescription: helpers.NewAttributeDescription("Forwarding router's address").String,
+										Required:            true,
+										Validators: []validator.String{
+											stringvalidator.RegexMatches(regexp.MustCompile(`(([0-9]|[1-9][0-9]|1[0-9][0-9]|2[0-4][0-9]|25[0-5])\.){3}([0-9]|[1-9][0-9]|1[0-9][0-9]|2[0-4][0-9]|25[0-5])(%[\p{N}\p{L}]+)?`), ""),
+											stringvalidator.RegexMatches(regexp.MustCompile(`[0-9\.]*`), ""),
+										},
+									},
+									"description": schema.StringAttribute{
+										MarkdownDescription: helpers.NewAttributeDescription("description of the static route").String,
+										Optional:            true,
+										Validators: []validator.String{
+											stringvalidator.LengthBetween(1, 30),
+										},
+									},
+									"tag": schema.Int64Attribute{
+										MarkdownDescription: helpers.NewAttributeDescription("Set tag for this route").AddIntegerRangeDescription(1, 4294967295).String,
+										Optional:            true,
+										Validators: []validator.Int64{
+											int64validator.Between(1, 4294967295),
+										},
+									},
+									"distance_metric": schema.Int64Attribute{
+										MarkdownDescription: helpers.NewAttributeDescription("Distance metric for this route").AddIntegerRangeDescription(1, 254).String,
+										Optional:            true,
+										Validators: []validator.Int64{
+											int64validator.Between(1, 254),
+										},
+									},
+									"permanent": schema.BoolAttribute{
+										MarkdownDescription: helpers.NewAttributeDescription("Permanent route").String,
+										Optional:            true,
+									},
+									"track": schema.StringAttribute{
+										MarkdownDescription: helpers.NewAttributeDescription("Enable object tracking for static route").String,
+										Optional:            true,
+										Validators: []validator.String{
+											stringvalidator.LengthBetween(1, 32),
+											stringvalidator.RegexMatches(regexp.MustCompile(`[\w\-\.:,_@#%$\+=\|;]+`), ""),
+										},
+									},
+									"metric": schema.Int64Attribute{
+										MarkdownDescription: helpers.NewAttributeDescription("Set metric for this route").AddIntegerRangeDescription(1, 16777214).String,
+										Optional:            true,
+										Validators: []validator.Int64{
+											int64validator.Between(1, 16777214),
+										},
+									},
+								},
 							},
 						},
 					},