+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - BREAKING CHANGE: Rename `iosxr_router_static` resource and data source to `iosxr_router_static_ipv4_unicast`
 - BREAKING CHANGE: Remove `iosxr_oc_system_config` resource and data source
 - Validate if referenced `device` exists in provider configuration
+- Add `vrfs`, `permanent`, `track` and `metric` attributes to `iosxr_router_static_ipv4_unicast` resource and data source
 
 ## 0.1.11